change module structure
